securegate-rbac/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pydantic
pytest
httpx
python-dotenv
